Display the number and status of each intervention event
Using a horizontal stack view, create and add styled circles
that are filled in if the event is completed. This implementation
has a known limitation, that it falls down if the number of
events results in views which are longer than the stack view.
Programatically applied constraints might work better, but we
can probably live with this for the demo.